dailybuild: suppress warnings
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7983 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWSkeletonSensor.cpp
@@ -49,7 +49,7 @@ void FWSkeletonSensor::Update() {
 				if (!bProcessed) {
 					// ‹ó‚«‚ª–³‚©‚Á‚½‚Ì‚Å’Ç‰Á‚·‚é
 					AddSkeleton();
-					ProcessSkeleton(sensor->GetSkeleton(i), skeletons.size()-1);
+					ProcessSkeleton(sensor->GetSkeleton(i), (int)skeletons.size()-1);
 				}
 			}
 		}
